refactor: stdout assertion & stderr assertion to handle failure cases better, add improved logs Move getting actual output from result, into the assertion.
